requirements.txt
pyTelegramBotAPI==4.5.0
python-dotenv==0.20.0
requests==2.27.1
telebot-calendar==1.2
loguru==0.6.0
geopy==2.2.0
validator-collection==1.5.0
peewee==3.14.10
python-telegram-bot-pagination==0.0.2